😆 Compleate the plugin base yooooo 🥰

+++ src/main/java/com/anas/intellij/plugins/ayah/settings/AyahSettingsState.java
@@ -62,7 +62,7 @@ public class AyahSettingsState implements PersistentStateComponent<AyahSettingsS
         this.basmalhOnStart = basmalhOnStart;
     }
 
-    public long getIntervalTimeBetweenNotifications() {
+    public int getIntervalTimeBetweenNotifications() {
         return intervalTimeBetweenNotifications;
     }
